requirements.txt
streamlit>=1.32.0
langchain>=0.1.13
langchain-community>=0.0.29
langchain-openai>=0.1.1
chromadb>=0.4.24
python-dotenv>=1.0.1
requests>=2.31.0
dashscope>=1.14.0
tiktoken>=0.6.0
unstructured>=0.12.5